requirements.txt
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jsonfy==0.4
MarkupSafe==3.0.2
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
urllib3==2.3.0
Werkzeug==3.1.3
